added processor interface and split piclassifier into it's own file
